+++ src/ezdxf/tools/text.py
@@ -1,6 +1,8 @@
 #  Copyright (c) 2021, Manfred Moitzi
 #  License: MIT License
-#  All tools in this module have to be independent from DXF entities!
+"""
+Tools in this module should be as independent from DXF entities as possible!
+"""
 import enum
 from typing import (
     List,
@@ -30,7 +32,7 @@ from ezdxf.lldxf.const import (
 )
 from ezdxf.math import Vec3, Vec2, Vertex
 from ezdxf.colors import rgb2int, RGB, int2rgb
-from .fonts import FontMeasurements, AbstractFont, FontFace
+from .fonts import FontMeasurements, AbstractFont, FontFace, make_font
 
 if TYPE_CHECKING:
     from ezdxf.eztypes import Text, MText, DXFEntity, Tags
@@ -1668,7 +1670,7 @@ def has_inline_formatting_codes(text: str) -> bool:
     )
 
 
-def is_upside_down_text_angle(angle: float, tol: float = 3.) -> bool:
+def is_upside_down_text_angle(angle: float, tol: float = 3.0) -> bool:
     """Returns ``True`` if the given text `angle` in degrees causes an upside
     down text in the :ref:`WCS`. The strict flip range is 90° < `angle` < 270°,
     the tolerance angle `tol` extends this range to: 90+tol < `angle` < 270-tol.
@@ -1679,11 +1681,11 @@ def is_upside_down_text_angle(angle: float, tol: float = 3.) -> bool:
         tol: tolerance range in which text flipping will be avoided
 
     """
-    angle %= 360.
-    return 90. + tol < angle < 270. - tol
+    angle %= 360.0
+    return 90.0 + tol < angle < 270.0 - tol
 
 
-def upright_text_angle(angle: float, tol: float = 3.) -> float:
+def upright_text_angle(angle: float, tol: float = 3.0) -> float:
     """Returns a readable (upright) text angle in the range `angle` <= 90+tol or
     `angle` >= 270-tol. The angle is normalized to [0, 360).
 
@@ -1693,5 +1695,100 @@ def upright_text_angle(angle: float, tol: float = 3.) -> float:
 
     """
     if is_upside_down_text_angle(angle, tol):
-        angle += 180.
-    return angle % 360.
+        angle += 180.0
+    return angle % 360.0
+
+
+def leading(cap_height: float, line_spacing: float = 1.0) -> float:
+    """Returns the distance from baseline to baseline.
+
+    Args:
+        cap_height: cap height of the line
+        line_spacing: line spacing factor as percentage of 3-on-5 spacing
+
+    """
+    # method "exact": 3-on-5 line spacing = 5/3 = 1.667
+    # method "at least" is not supported
+    return cap_height * 1.667 * line_spacing
+
+
+def estimate_mtext_extents(mtext: "MText") -> Tuple[float, float]:
+    """Estimate the width and height of a single column
+    :class:`~ezdxf.entities.MText` entity.
+
+    This function is faster than the :func:`~ezdxf.tools.text_size.mtext_size`
+    function, but the result is very inaccurate if inline codes are used or
+    line wrapping at the column border is involved!
+
+    This function uses the optional `Matplotlib` package if available.
+
+    Returns:
+        Tuple[width, height]
+
+    """
+
+    def _make_font():
+        cap_height: float = mtext.dxf.get_default("char_height")
+        doc = mtext.doc
+        if doc:
+            style = doc.styles.get(mtext.dxf.get_default("style"))
+            if style is not None:
+                return style.make_font(cap_height)  # type: ignore
+        return make_font(const.DEFAULT_TTF, cap_height=cap_height)
+
+    return estimate_mtext_content_extents(
+        content=mtext.text,
+        font=_make_font(),
+        column_width=mtext.dxf.get("width", 0.0),
+        line_spacing_factor=mtext.dxf.get_default("line_spacing_factor"),
+    )
+
+
+def estimate_mtext_content_extents(
+    content: str,
+    font: AbstractFont,
+    column_width: float = 0.0,
+    line_spacing_factor: float = 1.0,
+) -> Tuple[float, float]:
+    """Estimate the width and height of the :class:`~ezdxf.entities.MText`
+    content string. The result is very inaccurate if inline codes are used or
+    line wrapping at the column border is involved!
+    Column breaks ``\\N`` will be ignored.
+
+    This function uses the optional `Matplotlib` package if available.
+
+    Args:
+        content: the :class:`~ezdxf.entities.MText` content string
+        font: font abstraction based on :class:`ezdxf.tools.fonts.AbstractFont`
+        column_width: :attr:`MText.dxf.width` or 0.0 for an unrestricted column
+            width
+        line_spacing_factor: :attr:`MText.dxf.line_spacing_factor`
+
+    Returns:
+        Tuple[width, height]
+
+    """
+    max_width: float = 0.0
+    height: float = 0.0
+    cap_height: float = font.measurements.cap_height
+    has_column_width: bool = column_width > 0.0
+    lines: List[str] = fast_plain_mtext(content, split=True)  # type: ignore
+
+    if any(lines):  # has any non empty lines
+        line_count: int = 0
+        for line in lines:
+            line_width = font.text_width(line)
+            if has_column_width:
+                # naive line wrapping, does not care about line content
+                line_count += math.ceil(line_width / column_width)
+                line_width = min(line_width, column_width)
+            else:
+                line_count += 1
+            # Note: max_width can be smaller than the column_width, if all lines
+            # are shorter than column_width!
+            max_width = max(max_width, line_width)
+
+        spacing = leading(cap_height, line_spacing_factor) - cap_height
+        height = cap_height * line_count + spacing * (line_count - 1)
+
+    return max_width, height