+++ tests/test_02_dxf_graphics/test_200_line.py
@@ -4,9 +4,10 @@ import pytest
 import math
 
 from ezdxf.entities.line import Line
-from ezdxf.lldxf.const import DXF12, DXF2000
+from ezdxf.lldxf.const import DXF12, DXF2000, DXFValueError
 from ezdxf.lldxf.tagwriter import TagCollector, basic_tags_from_text
 from ezdxf.math import Matrix44
+from ezdxf.audit import Auditor
 
 TEST_CLASS = Line
 TEST_TYPE = "LINE"
@@ -211,6 +212,32 @@ def test_scaling():
     assert line.dxf.thickness == 4
 
 
+def test_copy_entity_transparency():
+    line = Line()
+    line2 = line.copy()
+    assert line2.dxf.hasattr("transparency") is False
+
+    line.transparency = 0.5
+    line2 = line.copy()
+    assert line2.dxf.transparency == 0x0200007F
+
+
+def test_setting_invalid_transparency_value_raises_exception():
+    line = Line()
+    with pytest.raises(DXFValueError):
+        line.dxf.transparency = 0
+
+
+def test_load_entity_with_invalid_transparency():
+    line = Line.from_text(ENTITY_INVALID_TRANSPARENCY)
+    # No auto fix in normal loading mode - Auditor fixes this issue at DXF
+    # attribute level!
+    assert line.dxf.transparency == 268435456
+    assert (
+        line.transparency == 0.0
+    ), "should replace invalid transparency by opaque"
+
+
 ERR_LINE = """0
 LINE
 5
@@ -227,6 +254,34 @@ AcDbLine
 1
 6
 Linetype
+10
+0.0
+20
+0.0
+30
+0.0
+11
+1.0
+21
+1.0
+31
+1.0
+"""
+
+ENTITY_INVALID_TRANSPARENCY = """0
+LINE
+5
+0
+330
+0
+100
+AcDbEntity
+8
+0
+440
+268435456
+100
+AcDbLine
 10
 0.0
 20