+++ tests/test_04_dxf_high_level_structs/test_424_audit.py
@@ -208,3 +208,15 @@ def test_remove_standalone_attrib_entities_from_blocks():
     auditor = doc.audit()
     assert len(list(msp)) == 0
     assert len(auditor.fixes) == 1
+
+
+def test_fix_invalid_transparency():
+    doc = ezdxf.new()
+    msp = doc.modelspace()
+    line = msp.add_line((0, 0), (1, 0))
+    # transparency value requires 0x02000000 bit set
+    line.dxf.unprotected_set("transparency", 0x10000000)
+    auditor = Auditor(doc)
+    line.audit(auditor)
+    assert line.dxf.hasattr("transparency") is False
+    assert len(auditor.fixes) == 1